Add testutils module and move 'DummyRtlSdr' into it
